# docs/api/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API - ERP Construção - RF-001",
    "description": "API para projeto na disciplina laboratório de inovação III - para servir aplicação web ERP.",
    "version": "1.0.0"
  },
  "paths": {
    "/produtos": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Listar Produtos",
        "description": "Lista os produtos ativos (ativo == true) — produtos deletados logicamente não aparecem.",
        "operationId": "listar_produtos_produtos_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProdutoResponse"
                  },
                  "type": "array",
                  "title": "Response Listar Produtos Produtos Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Produtos"
        ],
        "summary": "Criar Produto",
        "description": "Cria um novo produto e grava o usuário autenticado em criado_por.",
        "operationId": "criar_produto_produtos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/produtos/{produto_id}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "summary": "Buscar Produto",
        "description": "Busca um produto ativo específico pelo id. 404 se não existir ou estiver inativo.",
        "operationId": "buscar_produto_produtos__produto_id__get",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Produto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produtos"
        ],
        "summary": "Atualizar Produto",
        "description": "Atualiza um produto ativo e grava o usuário autenticado em atualizado_por.",
        "operationId": "atualizar_produto_produtos__produto_id__put",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Produto Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "summary": "Deletar Produto",
        "description": "Deleção lógica: NÃO remove a linha do banco (sem DROP/DELETE).\n\nApenas seta ativo = false e grava o usuário autenticado em deletado_por.",
        "operationId": "deletar_produto_produtos__produto_id__delete",
        "parameters": [
          {
            "name": "produto_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Produto Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProdutoCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome",
            "description": "Nome do produto"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao",
            "description": "Descrição detalhada do produto"
          },
          "preco": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Preco",
            "description": "Preço unitário do produto"
          },
          "unidade_medida": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Medida",
            "description": "Ex.: 'un', 'kg', 'm', 'm2', 'm3', 'saco'"
          },
          "quantidade_estoque": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantidade Estoque",
            "description": "Quantidade disponível em estoque",
            "default": 0
          },
          "categoria": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria",
            "description": "Categoria do produto"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "preco"
        ],
        "title": "ProdutoCreate",
        "description": "Payload esperado no POST /produtos."
      },
      "ProdutoResponse": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome",
            "description": "Nome do produto"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao",
            "description": "Descrição detalhada do produto"
          },
          "preco": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Preco",
            "description": "Preço unitário do produto"
          },
          "unidade_medida": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Medida",
            "description": "Ex.: 'un', 'kg', 'm', 'm2', 'm3', 'saco'"
          },
          "quantidade_estoque": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantidade Estoque",
            "description": "Quantidade disponível em estoque",
            "default": 0
          },
          "categoria": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria",
            "description": "Categoria do produto"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "criado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Por"
          },
          "atualizado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Por"
          },
          "deletado_por": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletado Por"
          },
          "criado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criado Em"
          },
          "atualizado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atualizado Em"
          },
          "deletado_em": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletado Em"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "preco",
          "id",
          "ativo"
        ],
        "title": "ProdutoResponse",
        "description": "Formato devolvido pela API em todas as rotas de produtos."
      },
      "ProdutoUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco"
          },
          "unidade_medida": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Unidade Medida"
          },
          "quantidade_estoque": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantidade Estoque"
          },
          "categoria": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria"
          }
        },
        "type": "object",
        "title": "ProdutoUpdate",
        "description": "Payload esperado no PUT /produtos/{id}.\n\nTodos os campos são opcionais para permitir atualização parcial\n(o cliente envia só o que quer alterar)."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}